[REF] test_point_of_sale: consolidate tests into test_point_of_sale_flow
