+++ addons/point_of_sale/tests/__init__.py
@@ -5,7 +5,6 @@ from . import test_point_of_sale_flow
 from . import test_frontend
 from . import test_point_of_sale_ui
 from . import test_anglo_saxon
-from . import test_point_of_sale
 from . import test_pos_controller
 from . import test_pos_cash_rounding
 from . import test_pos_setup
