+++ addons/point_of_sale/tests/test_point_of_sale_flow.py
@@ -1725,6 +1725,15 @@ class TestPointOfSaleFlow(TestPointOfSaleCommon):
 
         self.assertFalse(loaded_data['pos.config'][0]['pricelist_id'], False)
 
+        # merged similar test_no_default_pricelist test from TestPointOfSale
+        new_config = self.env["pos.config"].create({
+            "name": "usd config",
+            "available_pricelist_ids": [(6, 0, [self.public_pricelist.id])]
+        })
+        # make sure this doesn't pick a pricelist as default
+        self.assertEqual(new_config.pricelist_id, self.env['product.pricelist'],
+                         "POS config incorrectly has pricelist %s" % new_config.pricelist_id.display_name)
+
     def test_refund_rounding_backend(self):
         rouding_method = self.env['account.cash.rounding'].create({
             'name': 'Rounding up',
@@ -2030,6 +2039,36 @@ class TestPointOfSaleFlow(TestPointOfSaleCommon):
         product_form.type = "consu"
         product = product_form.save()
         self.assertFalse(product.combo_ids)
+
+    def test_product_combo_variants(self):
+        product_combo = self.env["product.combo"].create({
+            "name": "Magic Product combo",
+            "combo_item_ids": [
+                Command.create({
+                    "product_id": self.test_product3.id,
+                    "extra_price": 0,
+                }),
+            ],
+        })
+        # Add attribute and values, simulating variant creation
+        original_product_id = self.test_product3.id
+        product3_template = self.test_product3.product_tmpl_id
+        product3_template.with_context(create_product_product=True).write({
+            'attribute_line_ids': [(0, 0, {
+                'attribute_id': self.always_color_attribute.id,
+                'value_ids': [(6, 0, self.always_color_attribute.value_ids.ids)],
+            })],
+        })
+        # Check that original product should not be in combo anymore (replace by variants)
+        self.assertTrue(original_product_id not in product_combo.combo_item_ids.mapped('product_id').ids,
+                        "Original product should not be in combo")
+
+    def test_pos_bill_digits(self):
+        coin = self.env["pos.bill"].create({
+            "name": "0.005 not rounded",
+            "value": 0.005
+        })
+        self.assertEqual(coin.value, 0.005)
 
     def test_change_is_deducted_from_cash(self):
         self.pos_config.open_ui()
